requirements/base.txt
amqp==2.1.4
ansible==2.0.0.2
anyjson==0.3.3
billiard==3.5.0.2
celery==3.1.20
certifi==2016.9.26
cffi==1.6.0
coverage==4.3.1
Django==1.10.5
django-appconf==1.0.2
https://github.com/celery/django-celery/archive/v3.2.0.tar.gz
django-cleanup==0.4.2
django-cors-headers==1.3.1
django-crispy-forms==1.6.0
django-encrypted-fields==1.1.2
django-environ==0.4.1
django-extensions==1.7.5
django-fernet-fields==0.4
django-filter==1.0.1
django-measurement==2.4.0
django-memcache-status==1.3
django-modeltranslation==0.12
django-storages==1.5.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
drfdocs==0.0.11
ecdsa==0.13
et-xmlfile==1.0.1
Fabric==1.13.1
gunicorn==19.4.5
idna==2.1
jdcal==1.2
Jinja2==2.8.1
kombu==3.0.33
MarkupSafe==0.23
measurement==1.8.0
mpmath==0.19
newrelic==2.60.0.46
nose==1.3.7
numpy==1.11.1
olefile==0.43
opbeat==3.5.2
pandas==0.18.1
paramiko==1.16.0
Pillow==4.0.0
psycopg2==2.6.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyJWT==1.4.2
pytest==3.0.5
pytest-django==3.1.2
python-dateutil==2.5.3
python-keyczar==0.715
python-memcached==1.58
pytz==2015.7
PyYAML==3.11
redis==2.10.5
rest-pandas==0.4.0
simplejson==3.8.2
six==1.10.0
sqlparse==0.2.2
sympy==1.0
urllib3==1.19.1
viivakoodi==0.8.0
vine==1.1.3
virtualenv==15.1.0